requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
frozenlist==1.4.0
future==0.18.2
hubspot==0.1.14.dev0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
macholib==1.15.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
pip==23.2.1
requests==2.31.0
setuptools==58.0.4
six==1.15.0
tqdm==4.66.1
urllib3==2.0.4
python-dotenv==0.19.2
Werkzeug==2.3.7
wheel==0.37.0
yarl==1.9.2
zipp==3.16.2
gunicorn==20.1.0
openai
flask_sqlalchemy
psycopg2
sqlalchemy_utils
tiktoken
Flask-Migrate
flask-paginate